it works!

+++ src/Project/functions/tryRemoveOutput.ts
@@ -37,6 +37,9 @@ export function tryRemoveOutput(pathTranslator: PathTranslator, outPath: string)
 		// console.log("skipping " + outPath);
 		return;
 	}
+	if (outPath.endsWith(".meta")) {
+		return;
+	}
 	if (isOutputFileOrphaned(pathTranslator, outPath)) {
 		fs.removeSync(outPath);
 		LogService.writeLineIfVerbose(`remove ${outPath}`);
